+++ packages/asc-ui/src/utils/themeUtils.ts
@@ -132,17 +132,22 @@ export const outlineStyle = (
   outline-width: ${width}px;
 `
 
+/* we have chosen here to use a dubble selector '&&'.  
+ This will override a hover state with outlines.
+ introduced this when resolving issue: #131
+*/
 export const focusStyleOutline = (width?: number, offset?: number) => ({
   theme,
 }: {
   theme: Theme.ThemeInterface
-}) => css`
-  &&:focus {
-    ${outlineStyle(theme, width, offset)}
-  }
-`
+}) =>
+  css`
+    &&:focus {
+      ${outlineStyle(theme, width, offset)}
+    }
+  `
 
-export const focusStyleText = () => ({
+export const focusStyleFill = () => ({
   theme,
 }: {
   theme: Theme.ThemeInterface
@@ -151,6 +156,32 @@ export const focusStyleText = () => ({
     background-color: ${themeColor('support', 'focus')({ theme })};
   }
 `
+
+export enum FocusStyleEnum {
+  outline,
+  fill,
+  none,
+}
+
+/**
+ * @param  {keyoftypeofFocusStyleEnum='fill'} focusStyle
+ *
+ * decorates an element with one of the existing focus styles:
+ * - outline: draws a border around the element on focus
+ * - fill: fills the element background on focus
+ * - none: ignored the focus state
+ */
+export const getFocusStyle = (
+  focusStyle: keyof typeof FocusStyleEnum = 'fill',
+) => {
+  const styles = {
+    outline: focusStyleOutline(),
+    fill: focusStyleFill(),
+    none: '',
+  }
+
+  return styles[focusStyle]
+}
 
 /**
  * Util to hide the component for screen readers